Separating the components into smaller ones
