+++ src/components/Tab/tab.st.css
@@ -1,0 +1,2 @@
+/* tab.st.css CSS API */
+.root {}
