+++ src/components/TabPanel/tabPanel.st.css
@@ -1,0 +1,2 @@
+/* tabPanel.st.css CSS API */
+.root {}
