+++ src/components/Tabs/Tabs.tsx
@@ -1,52 +1,16 @@
 import { useTabListState } from "react-stately";
 import React, { useEffect, useRef, useState } from "react";
-import {
-  useTab,
-  useTabList,
-  useTabPanel,
-  useFocusRing,
-  mergeProps,
-} from "react-aria";
+import { useTabList, useFocusRing, mergeProps } from "react-aria";
+import Tab from "../Tab/Tab";
+import TabPanel from "../TabPanel/TabPanel";
 import { st, classes } from "./tabs.st.css";
-
-interface TabProps {
-  item: any;
-  state: any;
-  children: any;
-}
-
-function Tab({ item, state }: TabProps) {
-  const ref = useRef();
-  const { tabProps } = useTab(item, state, ref);
-
-  return (
-    <div {...tabProps} ref={ref}>
-      {item.rendered}
-    </div>
-  );
-}
-
-interface TabPanelProps {
-  state: any;
-  [key: string]: any;
-}
-
-function TabPanel({ state, ...props }: TabPanelProps) {
-  const ref = useRef();
-  const { tabPanelProps } = useTabPanel(props, state, ref);
-
-  return (
-    <div {...tabPanelProps} ref={ref}>
-      {state.selectedItem?.props.children}
-    </div>
-  );
-}
 
 export interface TabsProps {
   [key: string]: any;
+  className?: string;
 }
 
-export function Tabs(props: TabProps) {
+const Tabs = (props: TabsProps) => {
   const state = useTabListState(props);
   const ref = useRef();
   const { tabListProps } = useTabList(props, state, ref);
@@ -86,7 +50,7 @@ export function Tabs(props: TabProps) {
       <TabPanel key={state.selectedItem?.key} state={state} />
     </div>
   );
-}
+};
 
 Tabs.displayName = "Tabs";
